De-typed time fields + made message not array
The time field change was to let the date filter take care of them,
which I think expects a string.

The message thing was to change the name of the field containing a
message from the scheduler, so that it didn't conflict with the logstash
"message" field which contains the unaltered log line.