Made a good start on compute node syslog stashing.
There's still quite a few classes of message to either drop or save.

I've also started splitting the inputs, outputs, and filters. This is to
make it easier to test and deploy, since you can keep the paths for the
test files in inputs and test outputs in outputs on your test platform,
and then only commit the changes to the filters.

I could merge input and output into one file. I don't know. It probably
doesn't make much difference.